Move all integrate patterns into integrate.py
