+++ funsor/gaussian.py
@@ -8,7 +8,6 @@ from pyro.distributions.util import broadcast_shape
 import funsor.ops as ops
 from funsor.delta import Delta
 from funsor.domains import reals
-from funsor.integrate import Integrate
 from funsor.ops import AddOp, NegOp, SubOp
 from funsor.terms import Align, Binary, Funsor, FunsorMeta, Number, Slice, Subs, Unary, Variable, eager, reflect
 from funsor.torch import Tensor, align_tensor, align_tensors, materialize
@@ -565,50 +564,6 @@ def eager_neg(op, arg):
     info_vec = -arg.info_vec
     precision = -arg.precision
     return Gaussian(info_vec, precision, arg.inputs)
-
-
-@eager.register(Integrate, Gaussian, Variable, frozenset)
-def eager_integrate(log_measure, integrand, reduced_vars):
-    real_vars = frozenset(k for k in reduced_vars if log_measure.inputs[k].dtype == 'real')
-    if real_vars == frozenset([integrand.name]):
-        loc = cholesky_solve(log_measure.info_vec.unsqueeze(-1), log_measure._precision_chol).squeeze(-1)
-        data = loc * log_measure.log_normalizer.data.exp().unsqueeze(-1)
-        data = data.reshape(loc.shape[:-1] + integrand.output.shape)
-        inputs = OrderedDict((k, d) for k, d in log_measure.inputs.items() if d.dtype != 'real')
-        result = Tensor(data, inputs)
-        return result.reduce(ops.add, reduced_vars - real_vars)
-    return None  # defer to default implementation
-
-
-@eager.register(Integrate, Gaussian, Gaussian, frozenset)
-def eager_integrate(log_measure, integrand, reduced_vars):
-    real_vars = frozenset(k for k in reduced_vars if log_measure.inputs[k].dtype == 'real')
-    if real_vars:
-
-        lhs_reals = frozenset(k for k, d in log_measure.inputs.items() if d.dtype == 'real')
-        rhs_reals = frozenset(k for k, d in integrand.inputs.items() if d.dtype == 'real')
-        if lhs_reals == real_vars and rhs_reals <= real_vars:
-            inputs = OrderedDict((k, d) for t in (log_measure, integrand)
-                                 for k, d in t.inputs.items())
-            lhs_info_vec, lhs_precision = align_gaussian(inputs, log_measure)
-            rhs_info_vec, rhs_precision = align_gaussian(inputs, integrand)
-            lhs = Gaussian(lhs_info_vec, lhs_precision, inputs)
-
-            # Compute the expectation of a non-normalized quadratic form.
-            # See "The Matrix Cookbook" (November 15, 2012) ss. 8.2.2 eq. 380.
-            # http://www.math.uwaterloo.ca/~hwolkowi/matrixcookbook.pdf
-            norm = lhs.log_normalizer.data.exp()
-            lhs_cov = cholesky_inverse(lhs._precision_chol)
-            lhs_loc = cholesky_solve(lhs.info_vec.unsqueeze(-1), lhs._precision_chol).squeeze(-1)
-            vmv_term = _vv(lhs_loc, rhs_info_vec - 0.5 * _mv(rhs_precision, lhs_loc))
-            data = norm * (vmv_term - 0.5 * _trace_mm(rhs_precision, lhs_cov))
-            inputs = OrderedDict((k, d) for k, d in inputs.items() if k not in reduced_vars)
-            result = Tensor(data, inputs)
-            return result.reduce(ops.add, reduced_vars - real_vars)
-
-        raise NotImplementedError('TODO implement partial integration')
-
-    return None  # defer to default implementation
 
 
 __all__ = [
